# swagger/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Flipt REST API",
    "version": "0.0.0"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/flags/{key}": {
      "get": {
        "operationId": "GetFlag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/fliptFlag"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Flags"
        ]
      },
      "delete": {
        "operationId": "DeleteFlag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Flags"
        ]
      },
      "put": {
        "operationId": "UpdateFlag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/fliptFlag"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/fliptUpdateFlagRequest"
            }
          }
        ],
        "tags": [
          "Flags"
        ]
      }
    },
...
  },
  "definitions": {
...
    "fliptFlag": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean",
          "format": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "variants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/fliptVariant"
          }
        }
      }
    },
...
    "fliptUpdateFlagRequest": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
...
    "fliptVariant": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "flag_key": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}